rename : stockDetail (anon) 아래로 이동
