+++ app/(anon)/stock/[symbol]/components/StockInfo.tsx
