+++ app/(anon)/stock/[symbol]/components/SymbolLink.tsx
@@ -4,7 +4,7 @@ export const SymbolLink = ({ symbol }: { symbol: string }) => {
   const encodedSymbol = encodeURIComponent(symbol);
 
   return (
-    <Link href={`/user/stock/detail/${encodedSymbol}`}>
+    <Link href={`/(anon)/stock/detail/${encodedSymbol}`}>
       {symbol} 상세 페이지
     </Link>
   );
