+++ app/(anon)/stock/[symbol]/stock_detail.styled.ts
