+++ app/(anon)/stock/page.tsx
